fix(packaging): Tweaks to packaging to improve build and release process
