+++ package.json
@@ -26,10 +26,9 @@
     "prepush": "yarn test",
     "prelint-fix": "prettier --ignore-path .eslintignore --single-quote --write '**/*.{*{js,ts,tsx,jsx,json},babelrc,eslintrc,prettierrc,stylelintrc}'",
     "prerelease": "standard-version --prerelease rc",
-    "package": "cross-env PACKAGING=true yarn build && electron-builder build --config ./builderConfig.js --publish never && yarn post-pack",
-    "release": "cross-env PACKAGING=true yarn build && electron-builder build --config ./builderConfig.js --publish onTag && yarn post-pack",
-    "deploy": "cross-env PACKAGING=true yarn build && electron-builder build --config ./builderConfig.js --publish always && yarn post-pack",
-    "post-pack": "node ./post-pack.js",
+    "package": "cross-env PACKAGING=true yarn build && electron-builder build --config ./builderConfig.js --publish never ",
+    "release": "cross-env PACKAGING=true yarn build && electron-builder build --config ./builderConfig.js --publish onTag ",
+    "deploy": "cross-env PACKAGING=true yarn build && electron-builder build --config ./builderConfig.js --publish always ",
     "package-all": "yarn build && electron-builder build -mwl",
     "package-ci": "yarn postinstall && yarn build && electron-builder --publish always",
     "package-linux": "yarn build && electron-builder build --linux",
